Update ASAP.md
I just updated some spacing issues near the bottom of the document.  In the live web version there is no space between the text, "For these efforts, we are also reaching out to others within the community who are also creating data standards," included in the penultimate bullet point and the last bullet point.  I added a new line that I hope changes that on the web version.